Refactor file watching functionalities
